+++ __tests__/Validator.test.js
@@ -1,5 +1,5 @@
 const { ERROR_MESSAGE } = require('../src/constant/message');
-const Validator = require('../src/domain/Validator');
+const Validator = require('../src/js/domain/Validator');
 
 describe('숫자 입력 유효성 검사 테스트', () => {
   test('빈 값인 경우 예외가 발생한다.', () => {
