+++ src/js/constant/setting.js
