+++ src/js/domain/Comparer.js
