+++ src/js/domain/LottoMachine.js
