+++ src/js/domain/ProfitCaculator.js
