+++ src/js/domain/WinningLotto.js
