+++ src/js/step2-index.js
